+++ Application/Game/HorseDrawer.h
@@ -42,7 +42,7 @@ namespace YGame
 		/// </summary>
 		/// <param name="pParent"> : 親ポインタ</param>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		void Initialize(Transform* pParent, const uint32_t drawPriority) override;
+		void Initialize(Transform* pParent, const uint16_t drawPriority) override;
 
 		/// <summary>
 		/// 更新
@@ -69,14 +69,14 @@ namespace YGame
 		/// コンストラクタ
 		/// </summary>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		HorseDrawer(const uint32_t drawPriority);
+		HorseDrawer(const uint16_t drawPriority);
 
 		/// <summary>
 		/// コンストラクタ
 		/// </summary>
 		/// <param name="pParent"> : 親ポインタ</param>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		HorseDrawer(Transform * pParent, const uint32_t drawPriority);
+		HorseDrawer(Transform * pParent, const uint16_t drawPriority);
 
 		~HorseDrawer() = default;
