+++ Application/Game/PlayerDrawer.h
@@ -43,7 +43,7 @@ namespace YGame
 		/// </summary>
 		/// <param name="pParent"> : 親ポインタ</param>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		void Initialize(Transform* pParent, const uint32_t drawPriority) override;
+		void Initialize(Transform* pParent, const uint16_t drawPriority) override;
 
 		/// <summary>
 		/// 更新
@@ -70,14 +70,14 @@ namespace YGame
 		/// コンストラクタ
 		/// </summary>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		PlayerDrawer(const uint32_t drawPriority);
+		PlayerDrawer(const uint16_t drawPriority);
 
 		/// <summary>
 		/// コンストラクタ
 		/// </summary>
 		/// <param name="pParent"> : 親ポインタ</param>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		PlayerDrawer(Transform * pParent, const uint32_t drawPriority);
+		PlayerDrawer(Transform * pParent, const uint16_t drawPriority);
 
 		~PlayerDrawer() = default;
