+++ Application/Game/SlimeDrawer.h
@@ -39,7 +39,7 @@ namespace YGame
 		/// </summary>
 		/// <param name="pParent"> : 親ポインタ</param>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		void Initialize(Transform* pParent, const uint32_t drawPriority) override;
+		void Initialize(Transform* pParent, const uint16_t drawPriority) override;
 
 		/// <summary>
 		/// 更新
@@ -66,14 +66,14 @@ namespace YGame
 		/// コンストラクタ
 		/// </summary>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		SlimeDrawer(const uint32_t drawPriority);
+		SlimeDrawer(const uint16_t drawPriority);
 
 		/// <summary>
 		/// コンストラクタ
 		/// </summary>
 		/// <param name="pParent"> : 親ポインタ</param>
 		/// <param name="drawPriority"> : 描画優先度</param>
-		SlimeDrawer(Transform * pParent, const uint32_t drawPriority);
+		SlimeDrawer(Transform * pParent, const uint16_t drawPriority);
 
 		~SlimeDrawer() = default;
